fix entity for database, delete duplicate.

+++ src/main/java/lms/itcluster/confassistant/service/impl/UserServiceImpl.java
@@ -27,7 +27,7 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
 
     @Override
-    public User findById(int id) {
+    public User findById(long id) {
         return userRepository.findById(id).get();
     }
